pyproject.toml
[project]
name = "stream-bit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "boto3>=1.40.29",
    "cachetools>=6.2.0",
    "coverage>=7.10.6",
    "flask>=3.1.2",
    "pandas>=2.3.2",
    "pyarrow>=21.0.0",
    "pydantic>=2.11.7",
    "pytest>=8.4.2",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
